# Make simd8 64 uncopyable and other Visual Studio optimizations