--- a/singleheader/amalgamate_demo.cpp
+++ b/singleheader/amalgamate_demo.cpp
@@ -1,4 +1,4 @@
-/* auto-generated on Tue Jul  7 15:39:55 EDT 2020. Do not edit! */
+/* auto-generated on Mon 20 Jul 2020 18:53:18 EDT. Do not edit! */
 
 #include <iostream>
 #include "simdjson.h"