--- a/singleheader/amalgamate_demo.cpp
+++ b/singleheader/amalgamate_demo.cpp
@@ -1,4 +1,4 @@
-/* auto-generated on Tue 21 Jul 2020 15:17:46 EDT. Do not edit! */
+/* auto-generated on Tue 21 Jul 2020 15:20:03 EDT. Do not edit! */
 
 #include <iostream>
 #include "simdjson.h"