--- a/singleheader/simdjson.h
+++ b/singleheader/simdjson.h
@@ -1,4 +1,4 @@
-/* auto-generated on Tue 21 Jul 2020 15:17:46 EDT. Do not edit! */
+/* auto-generated on Tue 21 Jul 2020 15:20:03 EDT. Do not edit! */
 /* begin file include/simdjson.h */
 #ifndef SIMDJSON_H
 #define SIMDJSON_H