… cmake/simdjson-flags.cmake
… doc/performance.md
… src/generic/stage1/json_scanner.h
… src/generic/stage1/json_string_scanner.h
… src/generic/stage1/json_structural_indexer.h
… src/generic/stage1/utf8_fastvalidate_algorithm.h
… src/generic/stage1/utf8_lookup2_algorithm.h
… src/generic/stage1/utf8_lookup3_algorithm.h
… src/generic/stage1/utf8_lookup4_algorithm.h
… src/generic/stage1/utf8_lookup_algorithm.h
… src/generic/stage1/utf8_range_algorithm.h
… src/generic/stage1/utf8_zwegner_algorithm.h
… src/haswell/dom_parser_implementation.cpp
… src/haswell/simd.h
… src/westmere/dom_parser_implementation.cpp
… src/westmere/simd.h