This is a set of improvements to the code specific to Visual Studio to try to bridge the gap between Visual Studio and ClangCL.  Combined with an earlier commit, this makes Visual Studio 30% slower than ClangCL (under Visual Studio) while it was 1.9x slower a week ago.

Twitter.json:

| Platform | Overall | Stage 1 | Stage 2 |
|---|---|---|---|
| MSVC new | 2.0699 GB/s | 4.3916 GB/s | 4.7304 GB/s  |
| MSVC main branch [after commit](https://github.com/simdjson/simdjson/commit/d0ce2f0b5a6b93707bf6872cc591f32cf5d9eef2)  | 1.6424 GB/s| 2.9510 GB/s | 4.4473 GB/s |
| MSVC old (a week ago) | 1.4531 GB/s | 2.6907 GB/s | 3.6652 GB/s |
| ClangCL  | 2.7232 GB/s | 6.5038 GB/s | 5.7990 GB/s |

To reproduce:

```
git clone https://github.com/simdjson/simdjson.git
cd simdjson
mkdir build
cd build
cmake -DSIMDJSON_VISUAL_STUDIO_BUILD_WITH_DEBUG_INFO_FOR_PROFILING=ON ..
cmake --build . --target parse --config Release
./benchmark/Release/parse.exe ../jsonsexamples/twitter.json
```
It is possible to run `./benchmark/Release/parse.exe ../jsonsexamples/twitter.json` in VTune.

I suspect that more gains are possible but there are a few things that Visual Studio lacks ...

- Computed GOTOs
- likely/unlikely branch hints
- architecture-specific function attributes

So it may not be feasible to reach parity with ClangCL easily.


For x64 (linux)... Under GCC this is a small negative for stage 1 (penalty of ~1%) but it is a positive under clang for stage 1 (+ ~1%). It does not impact stage 2. Considering the large gain under Visual Studio, it is a net win all around. The results on ARM are inconclusive (seems all over the place though this might be a hardware issue).

This helps tremendously with issue https://github.com/simdjson/simdjson/issues/847


There remains a large gap between Visual Studio and ClangCL and, looking at the assembly, the diffference that jumps at you is the amount of vector/SIMD loads/store which is much higher with Visual Studio. 

Compiling with /arch:AVX2 does not help.


<img width="1078" alt="Screen Shot 2020-07-08 at 11 12 56 AM" src="https://user-images.githubusercontent.com/391987/86936801-7b435500-c10c-11ea-9030-e64bb6f4b2ea.png">
<img width="1072" alt="Screen Shot 2020-07-08 at 11 12 09 AM" src="https://user-images.githubusercontent.com/391987/86936811-7d0d1880-c10c-11ea-95ed-4e6d9e9f6e34.png">