@lemire 2020-07-08
Randomly adding __vectorcall to some of the functions that get inlined *may* have resulted in slightly better performance.