@lemire 2020-07-09
I have been trying to produce code examples...

```C++
struct go  {
    __m128i array[4];
};

__m128i orme(go & x) {
    __m128i z = x.array[0];
    for(size_t i = 1; i < 4; i++)
      z = _mm_or_si128(z, x.array[i]);
    return z;
}


```

Both GCC and clang compile it to...

```
orme(go&):
        movdqa  xmm0, XMMWORD PTR [rdi]
        movdqa  xmm1, XMMWORD PTR [rdi+32]
        por     xmm0, XMMWORD PTR [rdi+16]
        por     xmm1, XMMWORD PTR [rdi+48]
        por     xmm0, xmm1
        ret
```


Yet Visual Studio does the following...

```
__m128i orme(go &) PROC                    ; orme, COMDAT
        movdqu  xmm1, XMMWORD PTR [rcx]
        lea     rax, QWORD PTR [rcx+16]
        mov     ecx, 3
        npad    3
$LL4@orme:
        movdqu  xmm0, XMMWORD PTR [rax]
        lea     rax, QWORD PTR [rax+16]
        por     xmm0, xmm1
        movdqa  xmm1, xmm0
        sub     rcx, 1
        jne     SHORT $LL4@orme
        ret     0
```