@lemire 2020-07-09
Note that we have relatively few constants. All of the loads that we see are unlikely to be the compiler reloading constants. There may definitively be some register spilling, we have maybe 8 constants and two registers of inputs... but it should be quite modest...