@lemire 2020-07-09
Meanwhile, Visual Studio is perfectly able to produce sane code as long as SIMD registers are not involved... The following 

```C++
struct go {
    int array[4];
};

int orme(go & x) {
    int z = x.array[0];
    for(size_t i = 1; i < 4; i++)
      z |= x.array[i];
    return z;
}
```
compiles to 

```
        mov     edx, DWORD PTR [rcx+8]
        or      edx, DWORD PTR [rcx+12]
        or      edx, DWORD PTR [rcx+4]
        mov     eax, DWORD PTR [rcx]
        or      eax, edx
```

Why would Visual Studio treat SIMD registers as something that disables optimization opportunities?