@lemire 2020-07-14
This PR uses "pass-by-reference" in many instances, as opposed to "pass-by-value". Let us see why this is useful...

```C++
#ifdef _MSC_VER
#include <intrin.h>
#else
#include <x86intrin.h>
#endif
#include <cstdint>


struct simd8x32 {
 __m256i chunks;


};


bool inline is_ascii_ref(simd8x32& v) {
    return !_mm256_testz_si256(v.chunks, _mm256_set1_epi8(0b10000000));
}

bool inline is_ascii_copy(simd8x32 v) {
    return !_mm256_testz_si256(v.chunks, _mm256_set1_epi8(0b10000000));
}



simd8x32 * count_ascii_copy(simd8x32 * x) {
    while(is_ascii_copy(*x)) {
        x++;
    }
    return x;
}

simd8x32 * count_ascii_ref(simd8x32 * x) {
    while(is_ascii_ref(*x)) {
        x++;
    }
    return x;
}
```

The count_ascii_copy function gets compiled as 

```
simd8x32 * count_ascii_copy(simd8x32 *) PROC        ; count_ascii_copy, COMDAT
$LN13:
        push    rbp
        sub     rsp, 64                             ; 00000040H
        lea     rbp, QWORD PTR [rsp+32]
        and     rbp, -32                      ; ffffffffffffffe0H
        movups  xmm0, XMMWORD PTR [rcx]
        movups  xmm1, XMMWORD PTR [rcx+16]
        movaps  XMMWORD PTR $T1[rbp], xmm0
        movaps  XMMWORD PTR $T1[rbp+16], xmm1
        vmovdqu ymm0, YMMWORD PTR $T1[rbp]
        vptest  ymm0, YMMWORD PTR __ymm@8080808080808080808080808080808080808080808080808080808080808080
        vzeroupper
        je      SHORT $LN11@count_asci
        vmovdqu ymm2, YMMWORD PTR __ymm@8080808080808080808080808080808080808080808080808080808080808080
        npad    7
$LL2@count_asci:
        vmovups xmm0, XMMWORD PTR [rcx+32]
        add     rcx, 32                             ; 00000020H
        vmovaps XMMWORD PTR $T1[rbp], xmm0
        vmovups xmm1, XMMWORD PTR [rcx+16]
        vmovaps XMMWORD PTR $T1[rbp+16], xmm1
        vmovdqu ymm0, YMMWORD PTR $T1[rbp]
        vptest  ymm0, ymm2
        jne     SHORT $LL2@count_asci
$LN11@count_asci:
        mov     rax, rcx
        add     rsp, 64                             ; 00000040H
        pop     rbp
        ret     0
```

The count_ascii_ref function gets compiled as 
```
simd8x32 * count_ascii_ref(simd8x32 *) PROC   ; count_ascii_ref, COMDAT
        vmovdqu ymm1, YMMWORD PTR __ymm@8080808080808080808080808080808080808080808080808080808080808080
        vmovdqu ymm0, YMMWORD PTR [rcx]
        vptest  ymm0, ymm1
        je      SHORT $LN11@count_asci
        npad    13
$LL2@count_asci:
        vmovdqu ymm0, YMMWORD PTR [rcx+32]
        add     rcx, 32                             ; 00000020H
        vptest  ymm0, ymm1
        jne     SHORT $LL2@count_asci
$LN11@count_asci:
        mov     rax, rcx
        vzeroupper
        ret     0
```