@lemire 2020-07-14
But I would add that we can only take this so far. Doing everything by reference or by pointer is eventually going to tank the performance for other compilers.