@jkeiser 2020-07-17
As long as we use pass by reference exclusively in really_inline functions, I would expect it not to make any difference.