@lemire 2020-07-20
@jkeiser I have revisited the PR, making it less invasive (touching fewer lines of code). I took into account your comments.

I'll wait for this to pass our tests and I will merge.