@lemire 2020-07-21
Merging because I am tired of having this PR standing still.